Adding flit
